conditions to floods
