+++ src/containers/datasets/flood-protection/chart.tsx
@@ -2,7 +2,7 @@ import Legend from 'containers/legend';
 
 import Chart from 'components/chart';
 
-import type { Data } from '../types';
+import type { Data } from './types';
 
 const FloodProtectionChart = ({ data }: { data: Data }) => (
   <div className="grid w-full flex-1 grid-cols-3 items-center pb-10">
