+++ src/containers/map/index.tsx
@@ -132,7 +132,6 @@ const MapContainer = ({ mapId }: { mapId: string }) => {
     if (map) {
       setURLBounds(map.getBounds().toArray());
       setLocationBounds(null);
-      console.log(map.getStyle().layers);
     }
   }, [map, setURLBounds, setLocationBounds]);
